fix(ci): resolve all clippy findings so the clippy gate can pass
- clippy --fix for mechanical lints across the workspace
- iora-home: Box the TLS upstream variant, type aliases for complex row
  shapes, too_many_arguments allow on the internal download task,
  rustdoc list indentation fixes
- iora-shared: document the intentional from_str (not FromStr) name
- iora-files: type alias for subtree/copy futures
- dev-manager: allow the deliberately disabled QGA branch
- shared-config: fix doc list continuation

+++ iora-os/backend/services/iora-home/src/media_handler.rs
@@ -175,7 +175,7 @@ pub async fn continue_watching(
                 "series": item.get("SeriesName").and_then(|v| v.as_str()).unwrap_or("").to_string(),
                 "season": item.get("SeasonNumber").and_then(|v| v.as_i64()).unwrap_or(0),
                 "episode": item.get("EpisodeNumber").and_then(|v| v.as_i64()).unwrap_or(0),
-                "progress_percent": (progress as f64).round(),
+                "progress_percent": progress.round(),
                 "provider": "jellyfin",
                 // Proxied image — the media token stays server-side.
                 "image_url": if id.is_empty() { None } else {
@@ -219,7 +219,7 @@ pub async fn continue_watching(
                                 "series": item.get("grandparentTitle").and_then(|v| v.as_str()).unwrap_or("").to_string(),
                                 "season": item.get("parentIndex").and_then(|v| v.as_i64()).unwrap_or(0),
                                 "episode": item.get("index").and_then(|v| v.as_i64()).unwrap_or(0),
-                                "progress_percent": (progress as f64).round(),
+                                "progress_percent": progress.round(),
                                 "provider": "plex",
                             })
                         })
